Move Frontend interface to separate file
